backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-nmap==0.7.1
requests==2.32.3
